+++ lib/code_builder.dart
@@ -9,17 +9,27 @@ export 'src/matchers.dart' show equalsDart;
 export 'src/specs/annotation.dart' show Annotation, AnnotationBuilder;
 export 'src/specs/class.dart' show Class, ClassBuilder;
 export 'src/specs/code.dart' show Code, StaticCode, ScopedCode;
-export 'src/specs/code/expression.dart'
-    show
-        Expression,
-        literal,
-        literalNull,
-        literalBool,
-        literalTrue,
-        literalFalse;
 export 'src/specs/constructor.dart' show Constructor, ConstructorBuilder;
 export 'src/specs/directive.dart'
     show Directive, DirectiveType, DirectiveBuilder;
+export 'src/specs/expression.dart'
+    show
+        BinaryExpression,
+        CodeExpression,
+        Expression,
+        ExpressionEmitter,
+        ExpressionVisitor,
+        InvokeExpression,
+        InvokeExpressionType,
+        LiteralExpression,
+        LiteralListExpression,
+        literal,
+        literalNull,
+        literalBool,
+        literalList,
+        literalConstList,
+        literalTrue,
+        literalFalse;
 export 'src/specs/field.dart' show Field, FieldBuilder, FieldModifier;
 export 'src/specs/file.dart' show File, FileBuilder;
 export 'src/specs/method.dart'